return most recent genome build for the pipeline config file.
Brandon's original pipeline customization code replaced things
like %(genome|build)s with the path to the ELAND genome files.

What I did is made it possible to substitute keys like %(genome)s in
addition to %(genome|build)s. The idea is that the most config files
will be set to use whatever is the "most recent" build, but hopefully
at some point we'll provide some way of specifying which build.

The way I defined "most recent" genome build was to use the 
alphanum sort, that sorts mixed alpha/numeric strings in the 
'natural' order instead of ASCII order, thus "mm10" > "mm8".

For the genomes that we had installed right now this would work
for everything but arabadopsis--which appears to be using a version
number of MMDDYYYY. Though if we changed it to YYYYMMDD everything 
should work correctly.