make the mapped reads summary report more robust It's really useful for the summary report to report everything that mapped to the genome as a single entry, and everything that mapped to the contamination or spike-ins as seperate entries.
This version tags files that aren't symlined as 'last_dir_element/filename'
and files that are symlinked into the genome directory as 'filename'.
Since I have the spike-ins symlinked in, and the last element of the path
name is the genome name it makes it easy to use the 'last_dir_element' 
as the name to group all the per chromosome reads to.