Partially handle the changed Summary.htm file from the 0.3 version of the GAPipeline.
This update is incomplete as I'm pretty sure the xml serialization code
for the run xml file will break. However it does generate the summary
report for both the old summary file and the new post 0.3 file.

I also need to add unit tests for parsing and serializing the 0.3 
file format.