Utility to create srf files from a bustard directory
this version works, as long as you launch it in the bustard directory 
in question. There seems to be some messiness in the interaction between
how the list of arguments passed to Popen with shell=True has any file globs
expanded.

I had to switch from passing a list of arguments to Popen to string, 
and I'm still not sure if any of the code to try and change the directory 
to the bustard directory actually worked correctly. 

(which is why it only works when launching from the bustard directory)