Handle paired-end eland files.
This required changing the ELAND class to hold a list of dictionaries
from its previous implmentation where it was exporting an internal dictionary
of the lanes.

I decided to directly show the internal list and to remove the previous
dictionary methods to make it more obvious when code was expecting
the previous behavior.

Also a saved runfolder will now have eland files of the form 
s_<lane id>_<end id>.

Internally the end is 0 or 1, I tried to make the display show 1 or 2 for 
the users benefit though.