This is a partial merge of the stanford branch with the caltech branch of the web application, it doesn't work correctly yet, the libraries admin page is broken, and lacks the ability to browe the 'made_for' column.
This is based on a merge that started a few month ago, but I hadn't finished
the merge, I'll need to check for more updates from their branch soon.

During the process I decided it would be a good idea to update to django 1.0
which is going to make things even more unstable, so I thought I should
check this work in progess in before continuing.