Merge in Rami's changes from last friday.
Most of the admin pages work. Though there's a wsgi error with the reports. 
I'll try to figure out tomorrow.

the biggest difference between caltech trunk and stanford schemas right now
is caltech is using made_for as a foreign key, and stanford is using it
as a text field.