Try to make runfolder results extraction more robust If an IPAR or firecrest directory is missing some of the important matrix files that implies there isn't actually a valid run present, this patch will then (hopefully) issue a warning and skip that analysis run.
I also added an option to scripts/runfolder to allow a user to specify 
where the extracted results should go.

One questionable thing is that for one analysis some of the lanes 
were run as sequence and not an eland analysis so were I expected
all the lanes to have an eland genome, it doesn't for these.
I hope that the code doesn't lose the index after serializing and 
deserializing that chunk example.