Merge in the library list, detail, and results downloading feature from the Caltech live site.
There's several components in the frontend tree to render the pages, 
in addition this adds in some helper functions in pipelines.eland 
to simplify computing summary statistics for an eland lane.

I also needed to merge in a generator based makebed code for 
returning the files to the django database.

To use this, the settings file in this branch will need a variable 
RESULT_HOME_DIR to be set.