Make the public library browsing page support several features from django admin pages.
  * Search bar
  * Pagination
  * Filters

This took grabbing the "ChangeSet" class from django.contrib.admin and
slightly modifying, in addition to the css files.

To provide the css files I had to add the ability to serve static
files from the app level. I followed the django pattern that the
internal static pages would only be served by django when debug is true.

And because it's hard to split it out, I also added a new field
'hidden' to the library table. This allows hiding libraries from
the public library page (which is useful since a number of libraries
have a gel isolate pair, which many end users find rather confusing).