Consume output form the subprocesses
one of the times I was building the srf files the illumina2srf programs
stopped while the files were still incomplete, but with no CPU time.

It appears that the output from the child processes reached a point
where linux decided to block the process. This patch reads the output
from illumina2srf and if you run it with debug mode on it'll log it
otherwise it just ignores it.