Show the "Amplified from sample" as the parent library id and allow searching on that on the public library index.
Additionally detect if we have an eland_multi or eland extended file
and make the url to download those different files. (and change the
link name slightly to reflect you're getting a different file.)

finally don't bother showing the bed downloader for newer file types
as the bed convertor doesn't work on them.