Stop assuming that library_id == pk. I needed to fix the get_absolute_url to lookup by library_id, we needed a redirect function to get to the right admin page.
And for sequence runs I don't have any eland results so the get result files 
failed because it was assuming that the eland_result from the run xml would 
always have all 8 lanes defined in its dictionary. Now I check to see if its 
defined and if its not I don't display anything.