Watch more than one directory tree for modification. I also added code to record which directory watch, and what entry in that directory was being created (AKA whatever element of the runfolder is being touched, I record the root of the runfolder directory).
To support this I changed the config option in the benderjab file from
watchdir to watchdirs to make it a bit clearer that things have changed.

I still need to work out the communication protocol to copier so it
can figure out what to start copying. Also I'm recording the 
watchdirectory, but copier needs access to a different url. So there's
some question about where the right place to map watchdir to copy url source
might be.