Allow specifying the we're finished sequencing file in the config file. (Currently the default is netcopy_complete.txt, but that might not be a good choice).
While I was in there I changed it to catch the current event, and set a 
completed flag in the event instead of just directly sending the
sequencingFinished message.

I also put some effort into trying to only send startCopy messages for
things that look like runfolders, both by not sending a copy message
for the root of the watch directory and by checking to see if the 
top level directory that contains the event looks kind of like a runfolder
name. (AKA "[0-9]{6}_" )