Test that the email notification actually generates some emails.
I discovered that django when run in a test environment will store
email messages sent in django.core.mail.outbox, so I could make sure
that when the "run started" send email is clicked, it actually creates
some messages. (Though my current test, is just that the body is non-zero).