Put partial support back in for srf files.
Since I don't know if the srf file is supposed to be single or
paired end, this version assumes paired end unless you provide the
--single option.

Currently it'll give up if you try to convert a paired srf file
to a fastq file.

Also I made the code formatting in the make_parser function
look cleaner, and changed it to allow setting the logging verbosity
via command line options --verbose/--debug.