Create a class to handle mapping extension to ucsc view attributes.
This will attempt to guess if a library is paired end by seeing
if there are more paired end lanes than single read lanes.

The file exention matching was changed to use fnmatch instead of
ends with.

I still haven't added the ability to define extensions to ucsc DAF
view maps in a config file, as I don't know how to handle the case
of the .bam file that goes to a different view depending on if its
a paired end vs single end.

Also the ucsc_gather script is too long and parts of it need
to migrate into the rest of the htsworkflow tree.