Rename all the scripts to start with htsw-
In preparation for making a debian package I thought
I should try to limit my pollution of the command namespace.

so now everything that's actually an installed script starts
with htsw-. There's one left-over that still might be useful
in the future but isn't currently that was left with its old
name.