Add RDFa data to our library pages. While doing that I discovered that the library_index page was spitting up seriously bad HTML, with two DOCTYPE & html tags. It looks like it was a combination of my simplifications with the leftovers of brandon's "app" templates I ended up putting in two base.html templates.
I stripped out the app templates and tried to simplify the template
inheritance.

As a side effect the library detail page, now has a header and is using
a css based font.

There was also a minor issue where I discovered the paginator was showing
"X libraries libraries" because the views was adding "libraries" to
the changelist result count. (And the change list also adds in the name).

I needed to replace the default admin pagination template, as
RDFa uses xhtml, and xhtml doesn't support the &nbsp; entity.
(The solution is to use a unicode character for nbsp -- such as
&#160; -- unicode for non-breaking space.