Minor RDFa cleanups I changed the XMLSchema prefix to the example used by the W3c "xsd" instead of the random other sites example "xmls".
More importantly the content tag is content and not contents so
the places I was trying to replace the human readable value with
a standardized number or date were being ignored.

Lastly I figured out how to get django 1.1 to spit out a
XMLSchema#dateTime value. (which is not XMLSchema#datetime value)