Convert Rami's DataRun feature to something useful to us. (and more)
The data run captures some of the core values of a run, namely
the location of the result directory (relative to settings.RESULT_HOME_DIR)
and what the run cycle start/stop was.

The code will also scan the result directory looking for files whose
names match patterns stored in the FileType table and link those
to the DataRun in a DataFile table.

An unresolved issue is when & how often to scan the result directory --
currently when a live site flowcell is viewed it will look in the result
directory and import whatever it sees there. If it sees that result folder
again it wont bother to try and import it.

I might imagine problems if someone's hitting reload on a flowcell page
while the result files are being copied in.

I still need to implement a table to store results imported from the
run xml file. Also the old result folder scanning code used in the library
page should probably be updated to use the newer DataRun files.

Lastly there were some layout changes to a few pages and some
CSS rules got relocated.