Rework ucsc gather to use RDF models for gathering and storing track metadata.
I dumped all of my rdf namespace definitions into rdfhelp.py to comply
with don't-repeat-yourself.

The major step forward for this is I can now specify additional
attributes for a submission that aren't found in the htsworkflow database.

Also the daf model attribute lookup function implements a brute force
owl:sameAs search. (If it can't find the term you asked for, it'll search for
sameAs terms and look for those).

The TestDAFMapper unit test is a bit slow as it needs to check
the htsworkflow server for one of the test cases. I should come
up with a way to preload the required information into the testcase.