Further clean up ddf generation.
This still needs work as I ended up hard coding a sparql query
to support the submission I'm currently working on -- which is
unfortunate as the whole point of the push to RDF was to reduce
hard coding.

However it did simplify collecting information for make_ddf.

Using the query would also mean that the term copying I was doing
earlier, moving library attributes into each specific submission view
would be unecessary, since I can now easily query the graphs.

Probably what I need to do after the submission is to reduce the
term copying when importing a submisison directory, and add some
way of tying a sparql query to a specific imported daf.

Though I need to deal with the upcoming submission deadlines first.