Construct unit tests for inventory/hard disk long term storage component.
This involved RDFing a couple of the templates to make it easier
to make sure the right types of information were on the web pages.

I also added a new namespace http://jumpgate.caltech.edu/wiki/InventoryOntology#

There was also a missing </div> in one of the templates that I fixed.