Allow option to save/restore a sequence class to a RDF model. (After doing this I started having dreams of some set of mixins designed to persist data into different types of storage).
I also renamed the sql save to indicate that its going to a SQL
database.

Also I renamed one of my simplify Uris to stripNamespace
to make it clearer what it was actually doing.

simplify_uri just returns the end of a uri -- regardless of type.
stripNamespace removes a specific namespave from a uri.