Fix matching scanned sequence files to library IDs work for hiseq runs. The previous version was keying off of flowcell/lane so if you had multiple libraries from the same flowcell/lane all the sequences would end up in one of the libraries.
Hopefully this fixes that. Though to do this I ended up changing
the whole structure of condorfastq to be based on updating an RDF model.
This depends on the sequence.py module changes of saving things to
rdf models -- and the new code to infer library ids at that layer.