Make the inventory pages work with the new HTSChangeList.
There were a few problems, I was calling get_result_set in the template,
which requires the HTTP request which isn't available. I needed to be using
result_list instead.

I also changed the name in the context and failed to use the right model
for one of the indexes.