Generate manifest.txt files for submitting to ENCODE3.
Change trackhub generation from my previous template version to use
the Daler trackhub code.

This includes a feature to complain if you offer a submission set
name that diesn't exist.

Also the samples query returns all the submission components in
a single query instead of one of a time. (Which is much faster
way of doing things).