Flatten project hierarchy, moving djano applications out of htsworkflow.frontend into project root.
Unsuprisingly this also required updates to large chunks of
code to change the import paths for things.

I was motivated by the idea of trying to make my django applications more
portable. I tried some hacks with the python path to make it so they
could live in htsworkflow.appname, but django 1.7 seemed to want the app
to have only one import path, and this was the easiest way I could see to
solve that.