Convert samples app to use migrations.
A big problem is that 1.7 migrations cause problems
for the initial data fixture. For whatever reason the
fixture would run before the migration, unsurprisingly
causing trouble.

My choice was either to convert the fixture to a data
migration, or to find another way to generate test data.

Several sources recommended avoiding fixtures for test code
and suggested using something like "factory_boy" to use the
orm to create fake data. So I went ahead and tried an
implementation using that.