Delete temp objects to avoid constraint error
Django 1.10 with Postgres 9.? was throwing an error because the
samples_htsuser id that was assigned to an affiliation didn't exist in
the database.

The simplest solution I found was to delete the objects from the ORM,
I don't think they were ever actually created in the db so I'm not sure
how the constraint check was happening.